Updated node cache clear task

+++ directory/front/application/build/_actions/TaskClearNodeCache.php
@@ -14,12 +14,13 @@ class TaskClearNodeCache extends arch\task\Action {
     
     public function execute() {
         $this->io->write('Clearing node cache...');
-
         $path = $this->application->getLocalStoragePath().'/node';
-        $time = time();
 
-        core\io\Util::renameDir($path, 'node-'.$time);
-        core\io\Util::deleteDir($path.'/node-'.$time);
+        if(is_dir($path)) {
+            $time = time();
+            core\io\Util::renameDir($path, 'node-'.$time);
+            core\io\Util::deleteDir($path.'/node-'.$time);
+        }
         
         $this->io->writeLine(' done');
     }
